+++ ESPlatform.QRCode.IMS.Core/DTOs/MuaSamVatTu/Requests/CreatedSupplyRequest.cs
@@ -17,6 +17,6 @@ public class CreatedSupplyRequest
     [DisplayName("Xuất xứ")]
     public string XuatXu { get; set; } = string.Empty;
     
-    public int DonGia { get; set; }
-    public int SoLuong { get; set; }
+    public decimal DonGia { get; set; }
+    public decimal SoLuong { get; set; }
 }
